+++ domain/src/main/java/br/com/ministerio/recomeco/port/IRepository.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public interface IRepository<T> {
     List<T> listar();
-    T obterPorId(BigInteger id);
+    T obterPorId(Integer id);
     void criar(T t);
     T atualizar(T t);
-    void deletar(BigInteger id);
+    void deletar(Integer id);
 }
