+++ domain/src/main/java/br/com/ministerio/recomeco/service/Service.java
@@ -6,8 +6,8 @@ import java.util.List;
 public abstract class Service<T> {
     abstract List<T> listar();
     abstract List<T> listarPorNome(String nome);
-    abstract T obterPorId(BigInteger id);
+    abstract T obterPorId(Integer id);
     abstract void inserir(T t);
     abstract T atualizar(T t);
-    abstract void deletar(BigInteger id);
+    abstract void deletar(Integer id);
 }
